Fixed compiler warnings

+++ output_uint.c
@@ -6,7 +6,7 @@
 /*   By: ngontjar <ngontjar@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/06 19:21:53 by ngontjar          #+#    #+#             */
-/*   Updated: 2020/08/06 19:21:54 by ngontjar         ###   ########.fr       */
+/*   Updated: 2020/08/06 20:52:50 by ngontjar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ static void	init(size_t *len, int *prefix, int *w, t_data *flag)
 	{
 		flag->p = *len;
 	}
-	if (flag->width > flag->p)
+	if ((size_t)flag->width > flag->p)
 	{
 		*w = flag->width - flag->p - *prefix;
 	}
@@ -74,7 +74,7 @@ static void	justify_right(long long arg, const char *str, t_data *flag)
 {
 	int		w;
 	int		z;
-	int		len;
+	size_t	len;
 	int		prefix;
 
 	prefix = 2 * (flag->bit & FLAG_PREFIX && arg != 0);
